Changed function of 'subgit inspect' to print to stdout if no output file is specified

+++ subgit/inspect/git_inspect.py
@@ -12,8 +12,6 @@ from subgit.core import SubGit
 # 3rd party imports
 from ruamel import yaml
 
-import pysnooper
-
 
 log = logging.getLogger(__name__)
 
@@ -22,6 +20,7 @@ class GitInspect(SubGit):
     def __init__(self, config_file_name=None, answer_yes=None, is_archived=False):
         self.answer_yes = answer_yes
         self.is_archived = is_archived
+        self.config_file_name = config_file_name
 
         # Defaults config file to '.subgit-github.yml' if nothing is specified
         if not config_file_name:
@@ -58,7 +57,6 @@ class GitInspect(SubGit):
 
         return True
 
-    @pysnooper.snoop()
     def inspect_github(self, owner):
         """
         Given a username or organisation name, this method lists all repos connected to it
@@ -96,13 +94,6 @@ class GitInspect(SubGit):
             log.warning("No repos to write to file. Exiting...")
             return 1
 
-        if os.path.exists(self.subgit_config_file_path):
-            answer = self.yes_no(f"File: {self.subgit_config_file_path} already exists on disk, do you want to overwrite the file?")
-
-            if not answer:
-                log.error("Aborting inspection")
-                return 1
-
         for repo_name in sorted_names:
             repo_data = mapped_data[repo_name]
 
@@ -119,6 +110,30 @@ class GitInspect(SubGit):
                 "url": repo_data["sshUrl"],
             }
 
+        # If no config file name specfied, example file will be written to stdout
+        if not self.config_file_name:
+            log.info("Your '.subgit.yml' file would look like this:")
+            log.info("repos:")
+
+            for repo in repos:
+                branch = repos[repo]["revision"]["branch"]
+                url = repos[repo]["url"]
+                log.info(f"  {repo}:")
+                log.info(f"    revision:")
+                log.info(f"      branch: {branch}")
+                log.info(f"    url: {url}")
+            
+            return 0
+
+        # If config file name specified, inspect command will look for the file and ask for 
+        # confirmation whether user wants to overwrite this file
+        if os.path.exists(self.subgit_config_file_path):
+            answer = self.yes_no(f"File: {self.subgit_config_file_path} already exists on disk, do you want to overwrite the file?")
+
+            if not answer:
+                log.error("Aborting writing to file...")
+                return 1
+            
         yml = yaml.YAML()
         yml.indent(mapping=2, sequence=4, offset=2)
         with open(self.subgit_config_file_path, "w") as stream:
@@ -142,7 +157,7 @@ class GitInspect(SubGit):
                 "-o", "json",
                 "project", "list",
                 "--membership", "yes",
-                "--get-all",
+                "--all",
             ],
             shell=False,
             capture_output=True,
@@ -166,13 +181,6 @@ class GitInspect(SubGit):
             log.warning("No repos to write to file. Exiting...")
             return 1
 
-        if os.path.exists(self.subgit_config_file_path):
-            answer = self.yes_no(f"File: {self.subgit_config_file_path} already exists on disk, do you want to overwrite the file?")
-
-            if not answer:
-                log.error("Aborting inspection")
-                return 1
-
         for repo_name in sorted_names:
             repo_data = mapped_data[repo_name]
 
@@ -182,6 +190,30 @@ class GitInspect(SubGit):
                 },
                 "url": repo_data["ssh_url_to_repo"],
             }
+
+        # If no config file name specfied, example file will be written to stdout
+        if not self.config_file_name:
+            log.info("Your '.subgit.yml' file would look like this:")
+            log.info("repos:")
+
+            for repo in repos:
+                branch = repos[repo]["revision"]["branch"]
+                url = repos[repo]["url"]
+                log.info(f"  {repo}:")
+                log.info(f"    revision:")
+                log.info(f"      branch: {branch}")
+                log.info(f"    url: {url}")
+
+            return 0
+
+        # If config file name specified, inspect command will look for the file and ask for 
+        # confirmation whether user wants to overwrite this file
+        if os.path.exists(self.subgit_config_file_path):
+            answer = self.yes_no(f"File: {self.subgit_config_file_path} already exists on disk, do you want to overwrite the file?")
+
+            if not answer:
+                log.error("Aborting writing to file...")
+                return 1
 
         yml = yaml.YAML()
         yml.indent(mapping=2, sequence=4, offset=2)
